move vessel summary to activity
